refactor(agent): fix core-layer abstraction violations
agent/core's flow nodes bypassed the module's own fileService DI seam
and reached for the static AbsoluteFS class directly for output writes
and media-attachment existence checks — inconsistent with how every
other file operation in this module is injected via services.fileService
(TaskRunFileService). Add pathExists/writeFile/appendFile to
TaskRunFileService and route ResponseCycleFlow and ToolUseDispatchNode
through it instead.

Also close a real dependency inversion: agent/core/flows/CommonCycleTypes
imported formatPostCompactionContext from src/tools/delegation, the wrong
direction since src/tools is meant to consume agent/core. Move the
(pure, dependency-free) function to a new agent/core/flows/postCompactionContext
module and add @tools/** to agent/core's restricted-import patterns so
this can't silently recur.
