+++ src/agent/core/flows/ResponseCycleFlow.ts
@@ -1,5 +1,3 @@
-import { dirname } from 'node:path';
-
 import { z } from 'zod';
 
 import { BaseNode, Flow } from '@agent/node';
@@ -26,7 +24,6 @@ import { useLaunchRunContext } from '@agent/runtime/RunContext';
 import type { ToolDefinition } from '@model/ToolDefinition';
 import { MESSAGE_TYPES, AgentFileLocationSchema } from '@shared/schemas';
 import { OUTPUT_END_TAG } from '@shared/schemas/output';
-import { AbsoluteFS } from '@utils/files';
 import { toErrorMessage } from '@utils/errors/errorMessage';
 import { extractScratchpad } from '@utils/text/xmlExtraction';
 
@@ -110,9 +107,11 @@ class ResponsePrepNode<C> extends BaseNode<
   ResponseCycleServices<C>
 > {
   async prep(shared: ResponseCycleShared): Promise<ResponsePrepResult> {
-    const { prompt, userVarChannels, runScope } = this.services;
+    const { prompt, userVarChannels, runScope, fileService } = this.services;
     const interrupted = runScope.signal.aborted;
-    const exists = await AbsoluteFS.exists(shared.outputLocation.absolutePath);
+    const exists = await fileService.pathExists(
+      shared.outputLocation.absolutePath,
+    );
     const systemPrompt = interrupted
       ? undefined
       : await getSystemPromptWithRules(prompt.systemPrompt, {
@@ -337,7 +336,7 @@ class ResponseProcessNode<C> extends BaseNode<
     prepRes: ProcessPrepResult,
     execRes: ProcessNodeResult,
   ): Promise<string | undefined> {
-    const { round, workspace, logger } = this.services;
+    const { round, workspace, logger, fileService } = this.services;
     const modelHandler = this.services.modelCell.handler;
 
     if (execRes.kind === 'skipped') {
@@ -376,11 +375,9 @@ class ResponseProcessNode<C> extends BaseNode<
     const { outputLocation } = shared;
     const connector = result.bestConnector;
 
-    await AbsoluteFS.ensureDir(dirname(outputLocation.absolutePath));
-
     if (!shared.outputExists) {
       logger.debug(`Creating new file: ${outputLocation.absolutePath}`);
-      await AbsoluteFS.write(
+      await fileService.writeFile(
         outputLocation.absolutePath,
         result.processedResponse,
       );
@@ -389,7 +386,7 @@ class ResponseProcessNode<C> extends BaseNode<
       logger.debug(
         `Appending to existing file: ${outputLocation.absolutePath}`,
       );
-      await AbsoluteFS.appendFile(
+      await fileService.appendFile(
         outputLocation.absolutePath,
         connector + result.processedResponse,
       );
